js lint fix
